LayoutTests/imported/w3c/web-platform-tests/html/dom/documents/dom-tree-accessors/nameditem-03-expected.txt


FAIL applet elements are (mostly) gone assert_equals: expected (undefined) undefined but got (string) "test1"
